Let Demo use desktop resolution in full-screen Trying to work around bug #471 reported by Jon Daniel. Bug was about demo not working in full-screen.
Basically these days changing the monitor mode as Irrlicht still does to set a full-screen resolution is no longer recommended.
Support for that is getting worse, especially on X11. Some flat-screens don't like that anyway and not to mention it tends to mess up the user desktop in many systems. So for now let's just use the desktop resolution.

If people want to change resolutions in full-screen the typical way by now is to use a ScreenQuad (like example 27.PostProcessing) with a render-target-texture and then adapt the size of that.

Also many games these days don't use monitor-modes at all, but use some borderless full-sized window instead. But Irrlicht still lacks support for that at this time.